Formatted with black

+++ ExanteTaxCalculator/src/domain/buy_item.py
@@ -1,13 +1,15 @@
 import datetime
 
 from money import Money
+from decimal import Decimal
+
 
 class BuyItem:
-   def __init__(self, asset_name : str, amount : int, paid : Money, commission : Money, date : datetime.date, transaction_id: int):
-      self.asset_name = asset_name
-      self.amount = amount
-      self.paid = paid
-      self.commission = commission
-      self.date = date
-      self.asset_left = amount # if some amount was sold
-      self.transaction_id = transaction_id
+    def __init__(self, asset_name: str, amount: Decimal, paid: Money, commission: Money, date: datetime.date, transaction_id: int):
+        self.asset_name = asset_name
+        self.amount = amount
+        self.paid = paid
+        self.commission = commission
+        self.date = date
+        self.asset_left = amount  # if some amount was sold
+        self.transaction_id = transaction_id
